requirements.txt
--- Core Quant & Data Analysis ---

pandas==2.2.0
numpy>=1.24.0
scipy>=1.11.0
matplotlib>=3.8.0

--- Exchange & DeFi Integration ---

ccxt>=4.2.0
web3>=6.15.0

--- Configuration & Environment ---

PyYAML>=6.0.1
python-dotenv>=1.0.0
requests>=2.31.0

--- Testing & Quality Assurance ---

pytest>=8.0.0
pytest-mock>=3.12.0